🧬feat(operations): add variable set and modularize bodies
