--- src/operations/body.rs
@@ -1,88 +1,0 @@
-use std::fmt::{Debug, Formatter};
-
-use derive_more::Display;
-
-use crate::operations::MemoryMeasurable;
-
-#[derive(Clone, Display, PartialEq, Eq)]
-pub enum OperationBody {
-    #[display("CounterIncrease{_0}")]
-    CounterIncrease(CounterIncreaseBody),
-    #[display("Snapshot{_0}")]
-    Snapshot(SnapshotBody),
-}
-
-impl Debug for OperationBody {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{self}")
-    }
-}
-
-impl MemoryMeasurable for OperationBody {
-    fn size(&self) -> u64 {
-        match self {
-            OperationBody::CounterIncrease(body) => body.size(),
-            OperationBody::Snapshot(body) => body.size(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Display, PartialEq, Eq)]
-#[display("(delta={delta})")]
-pub struct CounterIncreaseBody {
-    pub delta: i64,
-}
-
-impl CounterIncreaseBody {
-    pub fn new(delta: i64) -> Self {
-        Self { delta }
-    }
-}
-
-impl MemoryMeasurable for CounterIncreaseBody {
-    fn size(&self) -> u64 {
-        size_of::<i64>() as u64
-    }
-}
-
-#[derive(Debug, Clone, Display, PartialEq, Eq)]
-#[display("(size:{})", data.len())]
-pub struct SnapshotBody {
-    pub data: Box<[u8]>,
-}
-
-impl SnapshotBody {
-    pub fn new(data: Box<[u8]>) -> Self {
-        Self { data }
-    }
-}
-
-impl MemoryMeasurable for SnapshotBody {
-    fn size(&self) -> u64 {
-        self.data.len() as u64
-    }
-}
-
-#[cfg(test)]
-mod tests_operation_body {
-    use tracing::info;
-
-    use crate::operations::{
-        MemoryMeasurable,
-        body::{CounterIncreaseBody, OperationBody},
-    };
-
-    #[test]
-    fn can_display_and_debug() {
-        let body = OperationBody::CounterIncrease(CounterIncreaseBody::new(123));
-        info!("{body} vs. {body:?}");
-        let s = format!("{body}");
-        assert!(s.starts_with("CounterIncrease(") && s.ends_with(')'));
-    }
-
-    #[test]
-    fn can_measure_body_size() {
-        let body = OperationBody::CounterIncrease(CounterIncreaseBody::new(123));
-        assert_eq!(body.size(), size_of::<i64>() as u64);
-    }
-}
