+++ src/operations/memory.rs
@@ -1,0 +1,3 @@
+pub trait MemoryMeasurable {
+    fn size(&self) -> u64;
+}
